Fixed ajax for updating name/leadership/points. More or less implemented leadership storing for squadrons. Added accessor and mutator for it in FleetShip model. Increased leadership column to mediumInteger

+++ database/migrations/2025_04_23_214021_create_fleet_ship_pivot_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->smallInteger('shields')->nullable();
             $table->string('armour')->nullable();
             $table->smallInteger('turrets')->nullable();
-            $table->smallInteger('leadership')->nullable();
+            $table->mediumInteger('leadership')->nullable();
             $table->smallInteger('squadron_counter')->nullable();
             $table->timestamps();
         });
